# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PTM-BMUP-Setting API",
        "description": "API untuk mengelola setting PTM-BMUP termasuk Core dan Member management",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development/Local Environment"
        },
        {
            "url": "https://staging-api.example.com",
            "description": "Staging Environment"
        },
        {
            "url": "https://api.example.com",
            "description": "Production Environment"
        }
    ],
    "paths": {
        "/api/setting/core": {
            "get": {
                "tags": [
                    "Core"
                ],
                "summary": "Get Core Settings",
                "description": "Mengambil data core settings",
                "operationId": "d77d586cd770f42aa9d32580300b3a3d",
                "responses": {
                    "200": {
                        "description": "Data core settings berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "PTM-BMUP"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Platform Training Management"
                                                },
                                                "logo": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/storage/logos/logo.png"
                                                },
                                                "created_by": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "updated_by": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Core"
                ],
                "summary": "Update Core Settings",
                "description": "Update core settings dengan file upload support",
                "operationId": "160a51f08479a585ed84add7c913af38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "PTM-BMUP Updated"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated Platform Training Management"
                                    },
                                    "logo": {
                                        "description": "Logo file upload",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "status_file": {
                                        "description": "0: no change, 1: replace/delete file",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Core settings berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Core updated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "PTM-BMUP Updated"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Updated Platform Training Management"
                                                },
                                                "logo": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/storage/logos/logo.png"
                                                },
                                                "created_by": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "updated_by": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "description": "JWT Token dari be-app-management project (httponly cookie)",
                "name": "token",
                "in": "cookie"
            },
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token untuk testing (opsional)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}